Test Client-Server

+++ 5-Offline/TodolistServer/src/id/ac/itb/todolist/server/ConnectionHandler.java
@@ -33,7 +33,7 @@ public class ConnectionHandler extends Thread {
             
             while (true) {
                 byte msgType = in.readByte();
-                int sessionId = -1;
+                long sessionId = -1;
                 
                 if (msgType != MSG_LOGIN) {
                     sessionId = in.readInt();
